server/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
dnspython==2.7.0; python_version >= '3.9'
email-validator==2.2.0; python_version >= '3.8'
fastapi[standard]==0.115.8; python_version >= '3.8'
fastapi-cli[standard]==0.0.7; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jinja2==3.1.5; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
psycopg2==2.9.10; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.20; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
rich==13.9.4; python_full_version >= '3.8.0'
rich-toolkit==0.13.2; python_version >= '3.8'
shellingham==1.5.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.45.3; python_version >= '3.9'
typer==0.15.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn[standard]==0.34.0; python_version >= '3.9'
watchfiles==1.0.4; python_version >= '3.9'
websockets==14.2; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
